[WIP][FEATURE] Initial PHP 7.2 images
Add `docker/php-official/7.2` directory which is
mostly a duplicate of `docker/php-official/7.1` and
generated Dockerfile.

Symlinked docker/php/7.2 to newly added on (see above).

Add `docker/php-dev/7.2` directory which is
mostly a duplicate of `docker/php-dev/7.1` and
generated Dockerfile.

Add `docker/php-apache/7.2` directory which is
mostly a duplicate of `docker/php-apache/7.1` and
generated Dockerfile.

Add `docker/php-apache-dev/7.2` directory which is
mostly a duplicate of `docker/php-apache-dev/7.1` and
generated Dockerfile.

Add `docker/php-nginx/7.2` directory which is
mostly a duplicate of `docker/php-nginx/7.1` and
generated Dockerfile.

All „Dockerfile“ files in this commit are auto generated
by running `./bin/console generate:dockerfile`.

Related: #211